feat: implement view story on map

+++ app/src/main/java/com/zuhal/storyapp/data/remote/models/Story.kt
@@ -22,8 +22,8 @@ data class Story(
     val description: String,
 
     @field:SerializedName("lon")
-    val lon: Float?,
+    val lon: Double?,
 
     @field:SerializedName("lat")
-    val lat: Float?
+    val lat: Double?
 ) : Parcelable
